MDBoostrap New Look

+++ src/main/java/com/intranet/controller/FileRestController.java
@@ -18,6 +18,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.ModelAndView;
+import org.springframework.web.servlet.view.RedirectView;
 
 import com.intranet.DemoApplication;
 import com.intranet.model.Encryptor;
@@ -34,6 +35,7 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.text.SimpleDateFormat;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Date;
 import java.util.HashMap;
@@ -52,7 +54,7 @@ import org.springframework.http.MediaType;
 public class FileRestController {
 
 	private final Logger log = LoggerFactory.getLogger(this.getClass());
-	
+
 	@Autowired
 	private UserService userService;
 
@@ -61,14 +63,14 @@ public class FileRestController {
 
 	@Autowired
 	private FolderService folderService;
-	
+
 	@RequestMapping(path = "/user/file/data/{id_file}", method = RequestMethod.GET)
 	public ResponseEntity<Map<String, Object>> autocompleteNames(@PathVariable("id_file") String nid){
 		Integer id = Integer.parseInt(Encryptor.decrypt(nid));
-		
+
 		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
 		User user = userService.findUserByEmail(auth.getName());
-		
+
 		HashMap<String, Object> map = null;
 		File file = fileService.findById(id);
 		try {
@@ -82,19 +84,19 @@ public class FileRestController {
 				datos.put("access", "" + file.getSharedUsers().size());
 				datos.put("creation", new SimpleDateFormat("dd-MM-yyyy hh:MM").format(file.getCreation()));
 				datos.put("update", new SimpleDateFormat("dd-MM-yyyy hh:MM").format(file.getLastUpdate()));
-				
+
 				if(file.getDownload() != null)
 					datos.put("download", new SimpleDateFormat("dd-MM-yyyy hh:MM").format(file.getDownload()));
 				else
 					datos.put("download", null);
-				
+
 				map.put(nid, datos);
 				return new ResponseEntity<Map<String, Object>>(map, HttpStatus.OK);
 			}
 		}catch(EntityNotFoundException e) {
 			log.error(e.getMessage());
 		}
-		
+
 		Folder folder = folderService.findById(id);
 		try {
 			if(user.getSharedFolders().contains(folder) || folder.getOwner().equals(user)) {
@@ -107,12 +109,12 @@ public class FileRestController {
 				datos.put("access", "" + folder.getSharedUsers().size());
 				datos.put("creation", new SimpleDateFormat("dd-MM-yyyy hh:MM").format(folder.getCreation()));
 				datos.put("update", new SimpleDateFormat("dd-MM-yyyy hh:MM").format(folder.getLastUpdate()));
-				
+
 				if(folder.getDownload() != null)
 					datos.put("download", new SimpleDateFormat("dd-MM-yyyy hh:MM").format(folder.getDownload()));
 				else
 					datos.put("download", null);
-				
+
 				map.put(nid, datos);
 				return new ResponseEntity<Map<String, Object>>(map, HttpStatus.OK);
 			}
@@ -124,10 +126,10 @@ public class FileRestController {
 			log.error("Incorrect file id");
 			return new ResponseEntity<Map<String, Object>>(HttpStatus.INTERNAL_SERVER_ERROR);
 		}
-		
+
 		return new ResponseEntity<Map<String, Object>>(HttpStatus.BAD_REQUEST);
 	}
-	
+
 	@RequestMapping(value="/user/file/{id_file}", method = RequestMethod.GET)
 	public ModelAndView embed(@PathVariable("id_file") String nid) {
 		Integer id = Integer.parseInt(Encryptor.decrypt(nid));
@@ -135,30 +137,53 @@ public class FileRestController {
 		ModelAndView modelAndView = new ModelAndView();
 		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
 		User user = userService.findUserByEmail(auth.getName());
-		File file = fileService.findById(id);
 
 		try {
+			File file = fileService.findById(id);
 			if(user.getSharedFiles().contains(file) || file.getOwner().equals(user)) {
-				java.io.File src = new java.io.File(getPath(file));
-				java.io.File temp = new java.io.File("src/main/resources/static/js/pdf/web/" + file.getName());
+				modelAndView.addObject("file", file);
+				
+				if(file.getFormat().equals("pdf")) {
+					java.io.File src = new java.io.File(getPath(file));
+					java.io.File temp = new java.io.File("src/main/resources/static/js/pdf/web/" + file.getName());
 
-				try {
-					Files.copy(Paths.get(src.getPath()) , Paths.get(temp.getPath()), java.nio.file.StandardCopyOption.REPLACE_EXISTING);
-				} catch (IOException e) {
-					log.error(e.getMessage());
+					try {
+						Files.copy(Paths.get(src.getPath()) , Paths.get(temp.getPath()), java.nio.file.StandardCopyOption.REPLACE_EXISTING);
+					} catch (IOException e) {
+						log.error(e.getMessage());
+					}
+					temp.delete();
+					
+					modelAndView.setViewName("user/file");
+					return modelAndView;
 				}
-
-				modelAndView.addObject("pdf", file.getName());
-				temp.delete();
+				
+				List<String> content = null;
+				try {
+					content = Files.readAllLines(Paths.get(getPath(file)));
+					List<String> lines = new ArrayList<String>();
+					for (String line : content) {
+						line = line.replace("\t", "&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;");
+						line = line.replace("<", "&#60;");
+						line = line.replace(">", "&#62;");
+						lines.add(line);
+					}
+					modelAndView.addObject("content", lines);
+					
+					modelAndView.setViewName("user/file");
+					return modelAndView;
+				} catch (IOException e) {
+					log.error("Reading not a plain text file");
+				}
 			}
 		}catch(EntityNotFoundException e) {
 			log.error(e.getMessage());
 		}
 
-		modelAndView.setViewName("user/file");
+		modelAndView.setView(new RedirectView("/user/files/" + user.getRoot().encrypt() , true));
 		return modelAndView;
 	}
-	
+
 	@PostMapping(path = "/user/edit/file/{id_file}", consumes = "text/plain")
 	public ResponseEntity<File> createFolder(@PathVariable("id_file") String nid, @RequestBody String name) {
 		Integer id = Integer.parseInt(Encryptor.decrypt(nid));
@@ -172,7 +197,7 @@ public class FileRestController {
 				File rename = new File();
 				rename.setParent(file.getParent());
 				rename.setName(name.concat("." + file.getFormat()));
-				
+
 				try {
 					java.io.File exist = new java.io.File(getPath(rename));
 					if(!exist.exists())
@@ -185,7 +210,7 @@ public class FileRestController {
 					log.error("Error renaming file");
 					return new ResponseEntity<File>(HttpStatus.CONFLICT);
 				}
-				
+
 				file.setName(name.concat("." + file.getFormat()));
 				fileService.update(file);
 				return new ResponseEntity<File>(HttpStatus.OK);
@@ -193,17 +218,17 @@ public class FileRestController {
 		} catch (EntityNotFoundException e) {
 			log.error(e.getMessage());
 		}
-		
+
 		Folder folder = folderService.findById(id);
 		try {
 			if(folder.getOwner().equals(user) || folder.getSharedUsers().contains(user)) {
 				Folder rename = new Folder();
 				rename.setParent(folder.getParent());
 				rename.setName(name);
-				
+
 				java.io.File oldD = new java.io.File(getPath(folder));
 				java.io.File newD = new java.io.File(getPath(rename));
-				
+
 				if(!newD.exists()) {
 					if(!oldD.renameTo(newD))
 						return new ResponseEntity<File>(HttpStatus.CONFLICT);
@@ -211,7 +236,7 @@ public class FileRestController {
 					log.error("Error renaming folder");
 					return new ResponseEntity<File>(HttpStatus.CONFLICT);
 				}
-				
+
 				folder.setName(name);
 				folderService.update(folder);
 				return new ResponseEntity<File>(HttpStatus.OK);
@@ -229,7 +254,7 @@ public class FileRestController {
 
 		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
 		User u = userService.findUserByEmail(auth.getName());
-		
+
 		if(emails.length != 0) {
 			for (int i = 0; i < emails.length; i++) {
 				emails[i] = emails[i].split("-")[1].trim();
@@ -241,7 +266,7 @@ public class FileRestController {
 					} catch (EntityNotFoundException e) {
 						log.error(e.getMessage());
 					}
-					
+
 					try {
 						Folder folder = folderService.findById(id);
 						shareFolder(folder, u, user);
@@ -252,17 +277,17 @@ public class FileRestController {
 			}
 			return new ResponseEntity<String>(HttpStatus.OK);
 		}
-			
+
 		return new ResponseEntity<String>(HttpStatus.BAD_REQUEST);
 	}
-	
+
 	@RequestMapping(value = "/user/file/shared/users/{id_resource}", method = RequestMethod.GET)
 	public HashMap<String, Object> sharedWith(@PathVariable("id_resource") String nid){
 		Integer id = Integer.parseInt(Encryptor.decrypt(nid));
 
 		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
 		User user = userService.findUserByEmail(auth.getName());
-		
+
 		File file = fileService.findById(id);
 		try {
 			if(file.getOwner().equals(user) || file.getSharedUsers().contains(user)) {
@@ -278,7 +303,7 @@ public class FileRestController {
 		} catch (EntityNotFoundException e) {
 			log.error(e.getMessage());
 		}
-		
+
 		Folder folder = folderService.findById(id);
 		try {
 			if(folder.getOwner().equals(user) || folder.getSharedUsers().contains(user)) {
@@ -294,10 +319,10 @@ public class FileRestController {
 		} catch (EntityNotFoundException e) {
 			log.error(e.getMessage());
 		}
-		
+
 		return new HashMap<String,Object>();
 	}
-	
+
 	@RequestMapping(value = "/user/upload/files/{id_folder}", method = RequestMethod.POST)
 	public ResponseEntity<File> uploadFileMulti(@RequestParam("files") MultipartFile[] uploadfiles, @PathVariable("id_folder") String nid) {
 		Integer id = Integer.parseInt(Encryptor.decrypt(nid));
@@ -326,7 +351,7 @@ public class FileRestController {
 		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
 		User user = userService.findUserByEmail(auth.getName());
 		File file = fileService.findById(id);
-		
+
 		try {
 			if(user.getSharedFiles().contains(file) || file.getOwner().equals(user)) {
 				Path pathFile = Paths.get(getPath(file));
@@ -345,10 +370,10 @@ public class FileRestController {
 				if(contentType == null) {
 					contentType = "application/octet-stream";
 				}
-				
+
 				file.setDownload(new Date());
 				fileService.update(file);
-				
+
 				return ResponseEntity.ok()
 						.contentType(MediaType.parseMediaType(contentType))
 						.header(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=\"" + resource.getFilename().replace("%20", " ")+ "\"")
@@ -367,18 +392,18 @@ public class FileRestController {
 			for (Folder f : folder.getFolders()) {
 				shareFolder(f, owner, user);
 			}
-			
+
 			for (File file : folder.getFiles()) {
 				shareFile(file, owner, user);
 			}
-			
+
 			folder.getSharedUsers().add(user);
 			user.getSharedFolders().add(folder);
 			folderService.update(folder);
 			userService.update(user);
 		}
 	}
-	
+
 	@Async
 	private void shareFile(File file, User owner, User user) {
 		if( (file != null && !file.getSharedUsers().contains(user) && !file.getOwner().equals(user)) && (file.getOwner().equals(owner) || file.getSharedUsers().contains(owner))) {
@@ -388,7 +413,7 @@ public class FileRestController {
 			userService.update(user);
 		}
 	}
-	
+
 	@Async
 	private void saveUploadedFiles(List<MultipartFile> files, int id) throws IOException {
 		Folder folder = folderService.findById(id);
@@ -397,9 +422,9 @@ public class FileRestController {
 
 		if(folder.getOwner().equals(user) || user.getSharedFolders().contains(folder)) {
 			for (MultipartFile file : files) {
-//				if (file.isEmpty()) {
-//					continue;
-//				}
+				//				if (file.isEmpty()) {
+				//					continue;
+				//				}
 
 				File f = new File();
 				f.setName(file.getOriginalFilename());
@@ -434,7 +459,7 @@ public class FileRestController {
 		path += file.getName();
 		return path;
 	}
-	
+
 	@Async
 	private String getPath(Folder folder) {
 		String path = "";
